added half stars cart can now add and take out items, and a temporary support

+++ backend/AtuhenticateApi/Models/Comment.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string UserId { get; set; }
         public int ProductId { get; set; }
-        public int Ratings { get; set; }
+        public double Ratings { get; set; }
         public string Comments { get; set; }
         public DateTime ReviewDate { get; set; }
